+++ src/Ext/Rules/Body.h
@@ -169,6 +169,10 @@ public:
 		Valueable<int> CombatLightDetailLevel;
 		Valueable<int> LightFlashAlphaImageDetailLevel;
 
+		Valueable<bool> MultiSelectNeutrals_Garrisonable;
+		Valueable<bool> MultiSelectNeutrals_TechBuilding;
+		Valueable<bool> MultiSelectNotOwned_TechBuilding;
+
 		Valueable<bool> EngineerAlwaysCaptureTech;
 
 		ExtData(RulesClass* OwnerObject) : Extension<RulesClass>(OwnerObject)
@@ -294,6 +298,10 @@ public:
 			, WarheadParticleAlphaImageIsLightFlash { false }
 			, CombatLightDetailLevel { 0 }
 			, LightFlashAlphaImageDetailLevel { 0 }
+
+			, MultiSelectNeutrals_Garrisonable { false }
+			, MultiSelectNeutrals_TechBuilding { false }
+			, MultiSelectNotOwned_TechBuilding { false }
 
 			, EngineerAlwaysCaptureTech { true }
 		{ }