+++ docs/Whats-New.md
@@ -497,6 +497,7 @@ New:
 - Auto load hotkey command (by psi-cmd, Aephiex)
 - Auto garrison hotkey command (by Aephiex)
 - Auto capture hotkey command (by Aephiex)
+- Allow limited multi selection of not owned objects to make better use of the auto load hotket (by Aephiex)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)